migration des templates

+++ prepare/variable.php
@@ -3,15 +3,19 @@
 class variable extends token {
     protected $nom = null;
 
-    function __construct($expression = null) {
+    function __construct($entree = null) {
         parent::__construct(array());
         
-        if (empty($expression)) {
+        if (empty($entree)) {
             $this->nom = null;
-        } elseif(count($expression) == 1) {
-          $this->nom = $expression[0];
-        } elseif(count($expression) == 2) {
-          $this->nom = $expression[1];
+        } elseif(count($entree) == 1) {
+            if ($entree[0]->checkClass('Token')) {
+                $this->nom = $entree[0]->getCode();
+            } else {
+                $this->nom = $entree[0];
+            }
+        } elseif(count($entree) == 2) {
+          $this->nom = $entree[1];
           $this->code = $this->nom->getCode();
         }
     }
@@ -25,7 +29,9 @@ class variable extends token {
     }
     
     function neutralise() {
-        $this->nom->detach();
+        if (is_object($this->nom)) {
+            $this->nom->detach();
+        }
     }
 
     function getRegex(){
